requirements.txt
fastapi==0.115.13
playwright==1.52.0
pydantic==2.11.7
scipy==1.15.3
torch==2.7.1
transformers==4.52.4
uvicorn
requests==2.32.4
tqdm==4.67.1
regex==2024.11.6
huggingface-hub==0.33.0
safetensors==0.5.3
tokenizers==0.21.1
openai>=1.0.0
numpy==1.26.4
python-dotenv==1.0.1